refactored themerevision item
